+++ src/Chirp.Core/Classes/Comment.cs
@@ -1,4 +1,3 @@
-
 namespace Chirp.Core.Classes;
 
 public class Comment
