+++ test/Chirp.Razor.Tests/Chirp.Tests.csproj
@@ -13,8 +13,6 @@
     <PackageReference Include="coverlet.collector" Version="6.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.8" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
-    <PackageReference Include="xunit" Version="2.5.3" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
   </ItemGroup>
 
   <ItemGroup>
@@ -26,10 +24,6 @@
     <ProjectReference Include="..\..\Chirp.Services\Chirp.Services.csproj" />
     <ProjectReference Include="..\..\src\Chirp.Core\Chirp.Core.csproj" />
     <ProjectReference Include="..\..\src\Chirp.Razor\Chirp.Razor.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Tests\" />
   </ItemGroup>
 
 </Project>
